fix: decimal type

+++ database/migrations/2025_09_17_101500_add_bilangan_and_bsu_bruto_to_produksi_detail_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('produksi_detail', function (Blueprint $table) {
-            $table->integer('bsu_bruto')->nullable()->after('pelaporan');
+            $table->decimal('bsu_bruto')->nullable()->after('pelaporan');
             $table->integer('bilangan')->nullable()->after('bsu_bruto');
         });
     }
